ai: refactor docker-compose.yml to use mariadb.
junie prompts:

update docker-compose.yml to rename wordpress-develop to wp-develop, mysql to wp-db, and prefix all other services with wp- as well.  hardwire use of mariadb 10 and php 8.4.  move all port mappings to a new docker-compose.override.sample.yml file.  remove all comments.  Update the README file to eliminate the section about apple silicon, and update references to mysql and mariadb as appropriate.  update .env.example to the new defaults as well.  update docs/junie-tasks.md to mark the relevant tasks as done.

fix references in bin/ to containers such as wordpress-develop to use the correct new container names

the wp-develop container is unable to start and logs the following error:
```
2025/04/18 12:43:09 [emerg] 1#1: host not found in upstream "php" in /etc/nginx/conf.d/default.conf:31
nginx: [emerg] host not found in upstream "php" in /etc/nginx/conf.d/default.conf:31
```

make a copy the default.conf from the container in tools/local-env/default.conf, edit it with the correct hostname, and add a volume mapping in docker-compose.yml to use this config