ai: add docs/plans/rebranding.md
junie prompts:
create docs/plans/rebranding.md which lays out the steps for renaming all instances of "WordPress" to "WigglePuppy" throughout the codebase.  There are a number of caveats to rebranding:

* copyright notices and author attributions should not change
* variable and function names should not change
* filenames and urls should not change
* mentions of WordPress as an organization (as opposed to the name of this software) should not change
* casing should be preserved, i.e. WordPress->WigglePuppy and wordpress->wigglepuppy

Make the following modifications:
* delete step 3 (Create Backup) from phase 1, as that is taken care of separately.
* As part of the inventory and analysis, categorize which replacement sites are amenable to using custom Rector rules for the replacement and which would be better as php scripts doing raw regex search/replace.
* Phase 2 should include development of custom Rectors for those replacements that can be automated (not the full implementation, just an some short sample snippets)

add a section to the doc about identifying and replacing other WordPress assets, specifically logo images, icons, etc.  replacement will be a manual process for those, but identifying those assets will still be a part of the inventorying step.